Create a basic Command structure
The AsyncCommand class is used to represent the execution of some action. Often the execution is triggered by some action in the UI. This is very important since it aims to separate the action from the data when used along side with ViewModels. It resembles the Action class from RxSwift: https://github.com/RxSwiftCommunity/Action, and the RACCommand class from ReactiveCocoa: http://codeblog.shape.dk/blog/2013/12/05/reactivecocoa-essentials-understanding-and-using-raccommand/.

This kind of pattern is very common int the context of MVVM in .NET Framework: https://msdn.microsoft.com/en-us/magazine/dn237302.aspx.