Using Dagger 2 on AndroidTest
A new test component was created in order to inject some mock objects used by specific components like the one used by the listing, this creates a kind of an integration test instead of an interface test. It's an interesting approach in which we will be able to test the UI behavior according to some input shuch as a request reponses.